+++ Viewer/FileChangeNotiThread.h
@@ -2,6 +2,7 @@
 
 #include <SThread.h>
 #include <SMutex.h>
+#include <atomic>
 #include "MainFrm.h"
 
 class FileChangeNotiThread : public SThread
@@ -10,6 +11,8 @@ public:
 	FileChangeNotiThread();
 	virtual ~FileChangeNotiThread(void);
 	SmpError fire(CMainFrame *pFrame, CString pathName);
+	void Stop();
+	unsigned Generation() const { return mGeneration.load(); }
 
 private:
 	int requestExitAndWait() { // do not call from outside, setup() will do
@@ -30,4 +33,5 @@ private:
 	CString mDirName;
 	CString mFileName;
 	SMutex mStateLock;
+	std::atomic<unsigned> mGeneration{0};
 };